--- .github/workflows/e2e.yml
@@ -1,64 +1,0 @@
-name: Tests and E2E
-
-on:
-  push:
-  workflow_dispatch:
-
-jobs:
-
-  e2e:
-    runs-on: ubuntu-latest
-    continue-on-error: true
-    timeout-minutes: 8
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-
-      - name: Setup Node 22
-        uses: actions/setup-node@v4
-        with:
-          node-version: 22
-
-      - name: Install Linux runtime dependencies for Electron
-        run: |
-          sudo apt-get update
-          sudo apt-get install -y \
-            xvfb \
-            libgtk-3-0 \
-            libnss3 \
-            libxss1 \
-            libasound2t64 \
-            libxtst6 \
-            libxrandr2 \
-            libgbm1 \
-            libxshmfence1
-
-      - name: Install app dependencies
-        run: npm install
-
-      - name: Fetch resource index (same as release build)
-        run: |
-          test -f src/index/index.sqlite
-          test -d src/index/resource_containers
-          rm -f src/index/resource_containers/en_ta-audio_vol2.tsrc
-          rm -f src/index/resource_containers/en_ta-checking_vol1.tsrc
-          rm -f src/index/resource_containers/en_ta-checking_vol2.tsrc
-          rm -f src/index/resource_containers/en_ta-gateway_vol3.tsrc
-          rm -f src/index/resource_containers/en_ta-intro_vol1.tsrc
-          rm -f src/index/resource_containers/en_ta-process_vol1.tsrc
-          rm -f src/index/resource_containers/en_ta-translate_vol1.tsrc
-          rm -f src/index/resource_containers/en_ta-translate_vol2.tsrc
-
-      - name: Bower install
-        run: npx bower install
-
-      - name: Run e2e tests
-        env:
-          DISPLAY: ":99"
-          # Chromium / main-process logs on stderr (renderer console often included when logging is on).
-          ELECTRON_ENABLE_LOGGING: "1"
-        run: |
-          Xvfb :99 -screen 0 1920x1080x24 &
-          # run-btt-e2e.sh tees Electron output here and to /tmp/btt-debug.log so the job log shows browser-side output before the Node e2e steps run.
-          bash scripts/e2e/run-btt-e2e.sh
