+++ e2e_tests/e2e/run-btt-e2e.sh
@@ -1,0 +1,22 @@
+#!/usr/bin/env bash
+
+set -euo pipefail
+
+REPO_ROOT="$(cd "$(dirname "${BASH_SOURCE[0]}")/../.." && pwd)"
+cd "$REPO_ROOT"
+
+# Stream main + renderer Chromium logs to CI stdout and keep a copy for the workflow artifact.
+DISPLAY="${DISPLAY:-:0}" ./node_modules/.bin/electron src/js/main.js \
+  --no-sandbox \
+  --remote-debugging-port=9222 \
+  --enable-logging \
+  2>&1 | tee /tmp/btt-debug.log &
+
+sleep 10s
+
+node --test e2e_tests/e2e/user-profile-setup.spec.mjs
+node --test e2e_tests/e2e/project-creation.spec.mjs
+
+sleep 3s
+
+node --test e2e_tests/e2e/project-navigation.spec.mjs
